static_cast no longer needed

+++ CorsixTH/Src/th_gfx_sdl.cpp
@@ -1114,7 +1114,7 @@ void sprite_sheet::draw_sprite(render_target* pCanvas, size_t iSprite, int iX,
     // converting them to rad
     uint32_t tick = gTickCount.currentTick();
     double currentVariation = 
-        sin(static_cast<double>(tick * 10 * M_PI / 180)) * 50;
+        sin(tick * 10 * M_PI / 180) * 50;
 
     // Vary from 40 to 80, so we set 60 + Variation
     err = SDL_SetTextureColorMod(pTexture, 0, 205 + currentVariation, 0);
